requirements.txt
# Epoch Trading System v2.0
# Core Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# UI Framework
PyQt6>=6.4.0
plotly>=5.18.0
matplotlib>=3.8.0
kaleido>=0.2.1  # For Plotly image export

# API & Data
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Time & Dates
pytz>=2023.3
pandas-market-calendars>=4.3.0

# AI Integration
anthropic>=0.18.0

# Development
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
